fix(ndt_scan_matcher): implicit double to float conversion

+++ localization/autoware_ndt_scan_matcher/include/autoware/ndt_scan_matcher/ndt_omp/multigrid_ndt_omp.h
@@ -158,7 +158,9 @@ public:
   inline void addTarget(const PointCloudTargetConstPtr & cloud, const std::string & target_id)
   {
     BaseRegType::setInputTarget(cloud);
-    target_cells_.setLeafSize(params_.resolution, params_.resolution, params_.resolution);
+    target_cells_.setLeafSize(
+      static_cast<float>(params_.resolution), static_cast<float>(params_.resolution),
+      static_cast<float>(params_.resolution));
     target_cells_.setInputCloudAndFilter(cloud, target_id);
   }
